refactor: move attributeTypes update into updateFileSettingsEffect
ref #2318
