+++ visualization/app/codeCharta/state/effects/updateFileSettings/updateFileSettings.effect.ts
@@ -5,8 +5,10 @@ import { isActionOfType } from "../../../util/reduxHelper"
 import { createEffect } from "../../angular-redux/effects/createEffect"
 import { Actions, ActionsToken } from "../../angular-redux/effects/effects.module"
 import { State } from "../../angular-redux/state"
+import { visibleFileStatesSelector } from "../../selectors/visibleFileStates.selector"
 import { FilesSelectionActions } from "../../store/files/files.actions"
 import { setState } from "../../store/state.actions"
+import { getMergedAttributeTypes } from "./utils/attributeTypes.merger"
 import { getMergedBlacklist } from "./utils/blacklist.merger"
 import { getMergedEdges } from "./utils/edges.merger"
 import { getMergedMarkedPackages } from "./utils/markedPackages.merger"
@@ -19,14 +21,16 @@ export class UpdateFileSettingsEffect {
 		this.actions$.pipe(
 			filter(action => isActionOfType(action.type, FilesSelectionActions)),
 			map(() => {
-				const { files } = this.state.getValue()
-				const visibleFiles = getVisibleFiles(files)
-				const withUpdatedPath = isPartialState(files)
+				const state = this.state.getValue()
+				const visibleFiles = getVisibleFiles(state.files)
+				const withUpdatedPath = isPartialState(state.files)
+				const allAttributeTypes = visibleFileStatesSelector(state).map(({ file }) => file.settings.fileSettings.attributeTypes)
 				return setState({
 					fileSettings: {
 						edges: getMergedEdges(visibleFiles, withUpdatedPath),
 						markedPackages: getMergedMarkedPackages(visibleFiles, withUpdatedPath),
-						blacklist: getMergedBlacklist(visibleFiles, withUpdatedPath)
+						blacklist: getMergedBlacklist(visibleFiles, withUpdatedPath),
+						attributeTypes: getMergedAttributeTypes(allAttributeTypes)
 					}
 				})
 			})
