+++ visualization/app/codeCharta/state/effects/updateFileSettings/utils/attributeTypes.merger.ts
@@ -1,4 +1,4 @@
-import { AttributeTypes } from "../../../codeCharta.model"
+import { AttributeTypes } from "../../../../codeCharta.model"
 
 export function getMergedAttributeTypes(allAttributeTypes: AttributeTypes[]): AttributeTypes {
 	const attributeTypesNodes = {}
